+++ src/Differ/ListDiffer.php
@@ -24,61 +24,12 @@ use InvalidArgumentException;
 class ListDiffer implements Differ {
 
 	/**
-	 * Use non-strict comparison and do not care about quantity.
-	 * This makes use of @see array_diff
-	 *
-	 * @since 0.4
-	 * @deprecated since 0.8, use new NativeArrayComparer() instead
-	 */
-	const MODE_NATIVE = 0;
-
-	/**
-	 * Use strict comparison and care about quantity.
-	 * This makes use of @see ListDiffer::strictDiff
-	 *
-	 * @since 0.4
-	 * @deprecated since 0.8, use null instead
-	 */
-	const MODE_STRICT = 1;
-
-	/**
 	 * @var ArrayComparer
 	 */
 	private $arrayComparer;
 
-	/**
-	 * @param ArrayComparer|int|null $arrayComparer Defaults to a StrictArrayComparer. The use of
-	 * the self::MODE_... constants is deprecated.
-	 *
-	 * The first argument is an ArrayComparer since version 0.8.
-	 * Before this it was an element of the ListDiffer::MODE_ enum.
-	 * The later is still supported though deprecated.
-	 */
-	public function __construct( $arrayComparer = null ) {
-		$this->arrayComparer = $this->getRealArrayComparer( $arrayComparer );
-	}
-
-	/**
-	 * @param ArrayComparer|int|null $arrayComparer Defaults to a StrictArrayComparer. The use of
-	 * the self::MODE_... constants is deprecated.
-	 *
-	 * @return ArrayComparer
-	 * @throws InvalidArgumentException
-	 */
-	private function getRealArrayComparer( $arrayComparer ): ArrayComparer {
-		if ( $arrayComparer === null || $arrayComparer === self::MODE_STRICT ) {
-			return new StrictArrayComparer();
-		}
-
-		if ( $arrayComparer === self::MODE_NATIVE ) {
-			return new NativeArrayComparer();
-		}
-
-		if ( is_object( $arrayComparer ) && $arrayComparer instanceof ArrayComparer ) {
-			return $arrayComparer;
-		}
-
-		throw new InvalidArgumentException( 'Got an invalid $arrayComparer' );
+	public function __construct( ArrayComparer $arrayComparer = null ) {
+		$this->arrayComparer = $arrayComparer ?? new StrictArrayComparer();
 	}
 
 	/**