+++ tests/phpunit/Differ/ListDifferTest.php
@@ -26,8 +26,6 @@ class ListDifferTest extends DiffTestCase {
 
 		return array(
 			'null' => array( null, $add ),
-			'native const' => array( ListDiffer::MODE_NATIVE, array() ),
-			'strict const' => array( ListDiffer::MODE_STRICT, $add ),
 			'native object' => array( new NativeArrayComparer(), array() ),
 			'strict object' => array( new StrictArrayComparer(), $add ),
 		);
@@ -40,11 +38,6 @@ class ListDifferTest extends DiffTestCase {
 		$differ = new ListDiffer( $arrayComparer );
 		$diff = $differ->doDiff( array( 1 ), array( 1, 1 ) );
 		$this->assertEquals( $expected, $diff );
-	}
-
-	public function testInvalidConstructorArgument() {
-		$this->expectException( 'InvalidArgumentException' );
-		new ListDiffer( 2 );
 	}
 
 	/**